MOD JavaCompletionTest: Added tests <method>.<field> and <method>.<method>
BUG JavaMethodParser: Removed wrong check in parseMethodArgumentList()

Reason: The expected result class must not be matched after the method and its argument list have been parsed, but after the whole expression starting with that method has been parsed. Example: If a String is expected, then getObject().toString() is fine. The method getObject() does not have to return a String.